==> code/01b_db_colombia_weekly_deaths_pop.R <==
rm(list=ls())
source("Code/00_functions.R")

# col <- read_xlsx("data_input/colombia/anexos-defunciones-covid-dept-semana-22-2022.xlsx",
#                  sheet = 2,
#                  skip = 11,
#                  col_types = rep("text", 15))

col <- read_xlsx("data_input/colombia/anexos-defunciones-covid-dept-semana-43-2022.xlsx",
                 sheet = 2,
                 skip = 11,
                 col_types = rep("text", 15))

pop_05_17 <- 
  read_xlsx("data_input/colombia/anexo-area-sexo-edad-proyecciones-poblacion-departamental_2005-2017.xlsx",
            skip = 11)

pop_18_50 <- 
  read_xlsx("data_input/colombia/anexo-proyecciones-poblacion-departamental_2018-2050.xlsx",
            skip = 11)


col2 <- 
  col %>% 
  rename(year = 1,
         geo = 2,
         week = 3,
         t_nat = 4,
         m_nat = 5,
         f_nat = 6,
         u_nat = 7,
         t_vio = 8,
         m_vio = 9,
         f_vio = 10,
         u_vio = 11,
         t_ees = 12,
         m_ees = 13,
         f_ees = 14,
         u_ees = 15) %>% 
  fill(year) %>% 
  fill(geo) %>% 
  drop_na(week) %>% 
  gather(-year, -geo, -week, key = type, value = dts) %>% 
  separate(type, c("sex", "cause"), sep = "_") %>% 
  filter(week != "Total") %>% 
  mutate(year = str_replace(year, "pr", ""),
         week = str_replace(week, "Semana ", ""),
         year = year %>% as.double(),
         week = week %>% as.double(),
         dts = dts %>% as.double()) %>% 
  filter(sex == "t") %>% 
  group_by(year, week, geo) %>% 
  summarise(dts = sum(dts)) %>% 
  ungroup()
  
unique(col2$year)
unique(col2$geo)
unique(col2$week)

col_sum <- 
  col2 %>% 
  group_by(geo, year) %>% 
  summarise(dts = sum(dts)) %>% 
  ungroup()

chunk <- 
  col2 %>% 
  filter(year == 2015,
         week == 1)

# re-scaling regions
dist_unk_geo <- 
  function(chunk){
    tot <- 
      chunk %>% 
      summarise(dts = sum(dts)) %>% 
      pull(dts)
    
    chunk %>% 
      filter(!geo %in% c("Sin información", "Extranjero")) %>% 
      mutate(dts = tot * dts/sum(dts))
  }

col3 <- 
  col2 %>% 
  group_by(year, week) %>% 
  do(dist_unk_geo(chunk = .data)) %>% 
  ungroup() %>% 
  mutate(isoweek = paste0(year, "-W", sprintf("%02d", week), "-7"),
         date = ISOweek2date(isoweek))

tot_nal <-
  col3 %>% 
  group_by(year, week, date, isoweek) %>% 
  summarise(dts = sum(dts)) %>% 
  ungroup() %>% 
  mutate(geo = "Total")

col4 <- 
  bind_rows(col3, tot_nal)


col4 %>% 
  filter(geo == "Total") %>% 
  ggplot()+
  geom_line(aes(date, dts))

col4 %>% 
  filter(geo == "Bogotá") %>% 
  ggplot()+
  geom_line(aes(date, dts))

unique(col4$geo) %>% sort

# population
# ~~~~~~~~~~

pop_05_17_2 <- 
  pop_05_17 %>% 
  rename(geo = 2,
         year = 3,
         area = 4) %>% 
  filter(area == "Total")

pop_18_50_2 <- 
  pop_18_50 %>% 
  rename(geo = 2,
         year = 3,
         area = 4) %>% 
  filter(area == "Total") %>% 
  mutate(year = year %>% as.double())

pop <- 
  bind_rows(pop_05_17_2,
            pop_18_50_2) %>% 
  select(-area) %>% 
  gather(-DP, -geo, -year, key = sex_age, value = pop) %>% 
  filter(!sex_age %in% c("Total Hombres", "Total Mujeres", "Total")) %>% 
  separate(sex_age, c("sex", "age"), sep = "_") %>% 
  mutate(geo = recode(geo,
                       "Archipiélago de San Andrés, Providencia y Santa Catalina" = "San Andrés y Providencia",
                       "Archipiélago de San Andrés" = "San Andrés y Providencia",
                       "Bogotá, D.C." = "Bogotá",
                       "Quindio" = "Quindío"))

pop_all <- 
  pop %>% 
  filter(sex == "Total") %>% 
  group_by(geo, year) %>% 
  summarise(pop = sum(pop)) %>% 
  ungroup() %>% 
  mutate(week = 26)

geos <- unique(pop$geo) %>% sort

pop_interpol <- 
  expand_grid(year = 2014:2022, week = 1:52, geo = geos) %>% 
  bind_rows(expand_grid(year = c(2015, 2020), week = 53, geo = geos)) %>% 
  arrange(geo, year, week) %>% 
  left_join(pop_all) %>% 
  group_by(geo) %>% 
  mutate(t = 1:n()) %>% 
  ungroup() %>% 
  group_by(geo) %>% 
  do(interpop(db = .data)) %>% 
  ungroup()

# visual inspection
# ~~~~~~~~~~~~~~~~~

pop_interpol %>% 
  filter(geo == "Bogotá") %>% 
  ggplot()+
  geom_line(aes(t, pop2))+
  geom_point(aes(t, pop), col = "red")
# weird case: San Andres

pop_interpol2 <- 
  pop_interpol %>% 
  select(-pop, -t) %>% 
  rename(pop = pop2)

pop_interpol3 <- 
  pop_interpol2 %>% 
  group_by(year, week) %>% 
  summarise(pop = sum(pop)) %>% 
  ungroup() %>% 
  mutate(geo = "Total") %>% 
  bind_rows(pop_interpol2)

# putting together weekly deaths and populations
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# standard codes
codes_std <- 
  read_csv("data_input/geo_codes_bra_col_mex.csv", 
           locale = readr::locale(encoding = "latin1")) %>% 
  filter(ISO_Code == "COL")

geo_codes <- 
  codes_std %>% 
  select(geo, geo_iso)

col_dts_pop <- 
  col4 %>% 
  left_join(pop_interpol3) %>% 
  mutate(geo = recode(geo,
                      "Atlántico" = "Atlantico",
                      "Bogotá" = "Bogota",
                      "Bolívar" = "Bolivar",
                      "Boyacá" = "Boyaca",
                      "Caquetá" = "Caqueta",
                      "Chocó" = "Choco",
                      "Córdoba" = "Cordoba",
                      "Guainía" = "Guainia",
                      "La Guajira" = "Guajira",
                      "Nariño" = "Narino",
                      "Quindío" = "Quindio",
                      "San Andrés y Providencia" = "San Andres",
                      "Valle del Cauca" = "Valle",
                      "Vaupés" = "Vaupes")) 

# grouping three regions
# grouping together regions in Colombia
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

reg_amazon <- c("Amazonas",
                "Caqueta",
                "Guainia",
                "Guaviare",
                "Putumayo",
                "Vaupes")

reg_orinoq <- c("Arauca", 
                "Casanare", 
                "Meta", 
                "Vichada")

reg_ejecaf <- c("Caldas", 
                "Risaralda", 
                "Quindio")

# reg_codes <- 
#   codes_std %>% 
#   select(geo = geo_gr, geo_iso = geo_iso_gr)

regs_dts_pop <- 
  col_dts_pop %>% 
  filter(geo %in% c(reg_amazon, reg_orinoq, reg_ejecaf)) %>% 
  mutate(geo = case_when(geo %in% reg_amazon ~ "Amazonia",
                         geo %in% reg_orinoq ~ "Orinoquia",
                         geo %in% reg_ejecaf ~ "Eje Cafetero",
                         TRUE ~ geo)) %>% 
  group_by(year, week, geo, isoweek, date) %>% 
  summarise(dts = sum(dts),
            pop = sum(pop)) %>% 
  ungroup()

col_out <- 
  bind_rows(col_dts_pop, regs_dts_pop) %>% 
  mutate(country = "COL") %>% 
  left_join(geo_codes)

write_rds(col_out, "data_inter/colombia_deaths_population_2015_2021.rds")
